# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Snake Arena API
  version: 1.0.0
  description: API for the Snake Arena game backend.
servers:
  - url: http://localhost:3000
    description: Local development server

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
        - createdAt

    LeaderboardEntry:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        username:
          type: string
        score:
          type: integer
        mode:
          type: string
          enum: [pass-through, walls]
        playedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - username
        - score
        - mode
        - playedAt

    ActivePlayer:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        currentScore:
          type: integer
        mode:
          type: string
          enum: [pass-through, walls]
        startedAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - currentScore
        - mode
        - startedAt

    Error:
      type: object
      properties:
        message:
          type: string
      required:
        - message

  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session_id

paths:
  /auth/login:
    post:
      summary: Login a user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/signup:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - username
                - email
                - password
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Validation error (e.g., duplicate email/username)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/logout:
    post:
      summary: Logout the current user
      tags: [Auth]
      responses:
        '200':
          description: Logout successful

  /auth/me:
    get:
      summary: Get current user profile
      tags: [Auth]
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated

  /leaderboard:
    get:
      summary: Get leaderboard entries
      tags: [Leaderboard]
      parameters:
        - in: query
          name: mode
          schema:
            type: string
            enum: [pass-through, walls]
          description: Filter by game mode
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: Number of entries to return
      responses:
        '200':
          description: List of leaderboard entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LeaderboardEntry'

    post:
      summary: Submit a new score
      tags: [Leaderboard]
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                score:
                  type: integer
                mode:
                  type: string
                  enum: [pass-through, walls]
              required:
                - score
                - mode
      responses:
        '201':
          description: Score submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  rank:
                    type: integer
                    nullable: true
        '401':
          description: Not authenticated

  /players:
    get:
      summary: Get active players
      tags: [Players]
      responses:
        '200':
          description: List of active players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivePlayer'

  /players/{playerId}:
    get:
      summary: Get player details
      tags: [Players]
      parameters:
        - in: path
          name: playerId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Player details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivePlayer'
        '404':
          description: Player not found